src/third_party/WebKit/LayoutTests/fast/dom/shadow/shadow-root-new-expected.txt

CONSOLE WARNING: line 19: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
This test checks the object 'host.createShadowRoot()' and internals.shadowRoot(div) are the same.

PASS root1 is root2
PASS successfullyParsed is true

TEST COMPLETE
